:truck: chore(api): move readmodel from shared to specific context
