+++ api/tests/tooling/domain-builder/factory/build-cpf-certification-result.js
@@ -1,4 +1,4 @@
-import { CpfCertificationResult } from '../../../../src/shared/domain/read-models/CpfCertificationResult.js';
+import { CpfCertificationResult } from '../../../../src/certification/session-management/domain/read-models/CpfCertificationResult.js';
 
 const buildCpfCertificationResult = function ({
   id = 1234,
